requirements.txt
pandas
graphviz
plotly
networkx
streamlit
matplotlib
timedelta
datetime
openpyxl
streamlit-elements
st-annotated-text
scikit-learn
seaborn
geopandas
joblib
xgboost
utils